requirements/scraper.txt
requests==2.28.2
scrapy==2.8.0
playwright==1.32.0
celery==5.2.7
redis==4.5.1
psycopg2-binary==2.9.10
SQLAlchemy==1.4.46